% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMPreprocessor.R
\name{\%||\%}
\alias{\%||\%}
\title{Null‑coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A value that may be `NULL`.}

\item{y}{A default value if `x` is `NULL`.}
}
\value{
`x` if not `NULL`, otherwise `y`.
}
\description{
Returns `y` if `x` is `NULL`, otherwise returns `x`.
}
