% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMPreprocessor.R
\name{or_or_operator}
\alias{or_or_operator}
\alias{\%||\%}
\title{%||% operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test for NULL.}

\item{y}{Default value if `x` is NULL.}
}
\value{
`x` if not NULL; otherwise `y`.
}
\description{
Null coalescing operator
}
\details{
Returns `y` if `x` is `NULL`, otherwise returns `x`.
}
